Fixed remaining build warnings. Fixed ShaderSignature for both D3D10 and D3D11. Resolves issue 659.

+++ source/xact3/Engine.cpp
@@ -97,7 +97,7 @@ namespace XACT3
 		params.file = file;
 		params.offset = offset;
 		params.flags = 0;
-		params.packetSize = packetSize;
+		params.packetSize = static_cast<WORD>(packetSize);
 
 		HRESULT hr = InternalPointer->CreateStreamingWaveBank(&params, &result);
 		if (RECORD_XACT3(hr).IsFailure)
